Player Healthbar Redesign (#85)

+++ src/main/java/org/davidfabio/game/Attackable.java
@@ -3,10 +3,10 @@ package org.davidfabio.game;
 import com.badlogic.gdx.graphics.Color;
 
 public interface Attackable {
-    public float getInitialHealth();
-    public void setInitialHealth(float newInitialHealth);
-    public float getHealth();
-    public void setHealth(float newHealth);
+    public int getInitialHealth();
+    public void setInitialHealth(int newInitialHealth);
+    public int getHealth();
+    public void setHealth(int newHealth);
     public boolean getIsActive();
     public void setIsActive(boolean newIsActive);
     public void setColor(Color color);
@@ -26,7 +26,7 @@ public interface Attackable {
     }
 
     public default void destroy(World world) {
-        this.setHealth(0f);
+        this.setHealth(0);
         this.setIsActive(false);
         this.playDestructionSound();
     }
