+++ src/main/java/org/davidfabio/game/EnemyBubble.java
@@ -6,7 +6,7 @@ public class EnemyBubble extends Enemy {
 
 
     @Override
-    public void init(float x, float y, float scale, float moveSpeed, float newInitialHealth, Color color) {
+    public void init(float x, float y, float scale, float moveSpeed, int newInitialHealth, Color color) {
         super.init(x, y, scale, moveSpeed, newInitialHealth, color);
 
         shape = new PolygonShape(64, scale);
