+++ src/main/java/org/davidfabio/game/EnemySpinner.java
@@ -11,7 +11,7 @@ public class EnemySpinner extends Enemy {
 
 
     @Override
-    public void init(float x, float y, float scale, float moveSpeed, float newInitialHealth, Color color) {
+    public void init(float x, float y, float scale, float moveSpeed, int newInitialHealth, Color color) {
         super.init(x, y, scale, moveSpeed, newInitialHealth, color);
 
         float[] vertices = new float[] {
