+++ src/main/java/org/davidfabio/game/EnemyTurret.java
@@ -8,7 +8,7 @@ public class EnemyTurret extends Enemy {
     BulletEnemySpawner bulletSpawner;
 
     @Override
-    public void init(float x, float y, float scale, float moveSpeed, float newInitialHealth, Color color) {
+    public void init(float x, float y, float scale, float moveSpeed, int newInitialHealth, Color color) {
         super.init(x, y, scale, moveSpeed, newInitialHealth, color);
 
         shape = new PolygonShape(4, scale);
